Refactor current sensor logic into multiple
In order to pave way for future sensors other than sleep, this breaks down the logic of OuraSensor into a base sensor and a sleep sensor. Logic that it's not related to the sensors has been extracted to helper modules. The former sensor logic remains only to import the other sensor files. Current changes does not include any significant changes or adds any new sensors. However, we're temporarily adding an optional type field with backwards compatibility.

This helps with #10
